chore: run tests on ssr fixture
