+++ packages/astro/test/fixtures/ssr-api-route/src/pages/custom-status.ts
